+++ lib/Migration/Version030200Date20260803233000.php
@@ -12,7 +12,7 @@ use OCP\Migration\SimpleMigrationStep;
 /**
  * Migrate the DB schema to Music v3.2.0 level from the v2.1.0 level
  */
-class Version030200Date20260729204000 extends SimpleMigrationStep {
+class Version030200Date20260803233000 extends SimpleMigrationStep {
 
 	/**
 	 * @param IOutput $output
@@ -38,6 +38,7 @@ class Version030200Date20260729204000 extends SimpleMigrationStep {
 		self::addUniqueIndexIfMissing($artists, 'user_hash_mbid_idx', ['user_id', 'hash', 'mbid']);
 
 		$albums = $schema->getTable('music_albums');
+		self::dropObsoleteColumn($albums, 'disk'); // migrated to oc_music_tracks in v0.13.1 in year 2020
 		self::addColumnIfMissing($albums, 'album_artist_uncertain', 'boolean', ['notnull' => false]);
 		// replace unique index for user_id/hash with one for user_id/hash/album_artist_id/mbid (the hash will no longer involve album_artist_id)
 		self::dropObsoleteIndex($albums, 'ma_user_id_hash_idx');
@@ -79,6 +80,15 @@ class Version030200Date20260729204000 extends SimpleMigrationStep {
 	private static function dropObsoleteIndex($table, string $name) : void {
 		if ($table->hasIndex($name)) {
 			$table->dropIndex($name);
+		}
+	}
+
+	/**
+	 * @param \Doctrine\DBAL\Schema\Table $table
+	 */
+	private static function dropObsoleteColumn($table, string $name) : void {
+		if ($table->hasColumn($name)) {
+			$table->dropColumn($name);
 		}
 	}
 }